Refactor main and create person_options, student_options and teacher_options to respect Single-responsiblity Principle
